Move id generation to be a method on HtlcParams
Currently we are passing a reference to an `HtlcParamas` structure to
`generate_id_*`.  The data is better encapsulated if we convert the
`generate_id_*` functions to be methods on `HtlcParams`.
