+++ cnd/src/swap_protocols/rfc003/mod.rs
@@ -31,19 +31,3 @@ pub use self::{
     secret::{FromErr, Secret, SecretHash},
     secret_source::*,
 };
-
-use crypto::{digest::Digest, sha2::Sha256};
-
-/// Generates a deterministic identifier from the secret hash and prefix.
-/// Returns SHA-256(prefix + secret_hash).
-pub fn generate_identifier(secret_hash: &SecretHash, prefix: &str) -> String {
-    let mut msg = String::from(prefix);
-    msg.push_str(&secret_hash.to_string());
-    hash(&msg)
-}
-
-fn hash(msg: &str) -> String {
-    let mut sha = Sha256::new();
-    sha.input_str(msg);
-    sha.result_str()
-}
