+++ cnd/src/swap_protocols/rfc003/state_machine.rs
@@ -12,6 +12,7 @@ use crate::swap_protocols::{
     },
     HashFunction, Timestamp,
 };
+use crypto::{digest::Digest, sha2::Sha256};
 use either::Either;
 use futures::{future, try_ready, Async, Future};
 use state_machine_future::{RentToOwn, StateMachineFuture};
@@ -55,6 +56,35 @@ impl<L: Ledger, A: Asset> HtlcParams<L, A> {
             secret_hash: request.secret_hash,
         }
     }
+
+    pub fn query_id_deployed(&self) -> String {
+        generate_identifier(&self.secret_hash, "deployed")
+    }
+
+    pub fn query_id_funded(&self) -> String {
+        generate_identifier(&self.secret_hash, "funded")
+    }
+
+    pub fn query_id_redeemed(&self) -> String {
+        generate_identifier(&self.secret_hash, "redeemed")
+    }
+
+    pub fn query_id_refunded(&self) -> String {
+        generate_identifier(&self.secret_hash, "refunded")
+    }
+}
+
+// Returns SHA-256(prefix + secret_hash).
+fn generate_identifier(secret_hash: &SecretHash, prefix: &str) -> String {
+    let mut msg = String::from(prefix);
+    msg.push_str(&secret_hash.to_string());
+    hash(&msg)
+}
+
+fn hash(msg: &str) -> String {
+    let mut sha = Sha256::new();
+    sha.input_str(msg);
+    sha.result_str()
 }
 
 #[derive(Debug, Clone, PartialEq)]
